updated function

+++ gfx_draw_line.c
@@ -21,12 +21,12 @@ void				gfx_draw_line(SDL_Surface *surf, Uint32 color, t_vector v1, t_vector v2)
 	int			overflow_x;
 	int			overflow_y;
 
-	cath_x = (v2.y - v1.y) < 0 ? (v2.y - v1.y) * -1 : (v2.y - v1.y);
-	cath_y = (v2.x - v1.x) < 0 ? (v2.x - v1.x) * -1 : (v2.x - v1.x);
+	cath_x = ((int)v2.y - (int)v1.y) < 0 ? ((int)v2.y - (int)v1.y) * -1 : ((int)v2.y - (int)v1.y);
+	cath_y = ((int)v2.x - (int)v1.x) < 0 ? ((int)v2.x - (int)v1.x) * -1 : ((int)v2.x - (int)v1.x);
 	overflow_y = cath_y - cath_x;
 	x = v1.x;
 	y = v1.y;
-	while (x != v2.x || y != v2.y)
+	while (x != (int)v2.x || y != (int)v2.y)
 	{
 		if (!(x < 0 || x >= surf->w || y < 0 || y >= surf->h))
 			set_pixel(surf, x, y, color);
@@ -34,12 +34,12 @@ void				gfx_draw_line(SDL_Surface *surf, Uint32 color, t_vector v1, t_vector v2)
 		if (overflow_x > -(cath_x))
 		{
 			overflow_y -= cath_x;
-			x += x < v2.x ? 1 : -1;
+			x += x < (int)v2.x ? 1 : -1;
 		}
 		else if (overflow_x < cath_x)
 		{
 			overflow_y += cath_y;
-			y += y < v2.y ? 1 : -1;
+			y += y < (int)v2.y ? 1 : -1;
 		}
 	}
 }
